+++ Pika/Pika.vcxproj
@@ -156,21 +156,25 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="src\Pika.h" />
-    <ClInclude Include="src\Pika\Application.h" />
-    <ClInclude Include="src\Pika\Core.h" />
-    <ClInclude Include="src\Pika\EntryPoint.h" />
+    <ClInclude Include="src\Pika\Core\Application.h" />
+    <ClInclude Include="src\Pika\Core\Core.h" />
+    <ClInclude Include="src\Pika\Core\EntryPoint.h" />
+    <ClInclude Include="src\Pika\Core\Layer.h" />
+    <ClInclude Include="src\Pika\Core\LayerStack.h" />
+    <ClInclude Include="src\Pika\Core\Log.h" />
+    <ClInclude Include="src\Pika\Core\Window.h" />
     <ClInclude Include="src\Pika\Events\ApplicationEvent.h" />
     <ClInclude Include="src\Pika\Events\Event.h" />
     <ClInclude Include="src\Pika\Events\KeyboardEvent.h" />
     <ClInclude Include="src\Pika\Events\MouseEvent.h" />
-    <ClInclude Include="src\Pika\Log.h" />
     <ClInclude Include="src\Pika\Platform\Windows\WindowsWindow.h" />
-    <ClInclude Include="src\Pika\Window.h" />
     <ClInclude Include="src\pkpch.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="src\Pika\Application.cpp" />
-    <ClCompile Include="src\Pika\Log.cpp" />
+    <ClCompile Include="src\Pika\Core\Application.cpp" />
+    <ClCompile Include="src\Pika\Core\Layer.cpp" />
+    <ClCompile Include="src\Pika\Core\LayerStack.cpp" />
+    <ClCompile Include="src\Pika\Core\Log.cpp" />
     <ClCompile Include="src\Pika\Platform\Windows\WindowsWindow.cpp" />
     <ClCompile Include="src\pkpch.cpp">
       <PrecompiledHeader>Create</PrecompiledHeader>
