+++ Pika/src/Pika/Core/Layer.cpp
@@ -1,0 +1,7 @@
+#include "pkpch.h"
+#include "Layer.h"
+
+Pika::Layer::Layer(const std::string& vLayerName) :
+	m_DebugName{ vLayerName }
+{
+}
