+++ Pika/src/Pika/Events/Event.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "pkpch.h"
-#include "Pika/Core.h"
+#include "Pika/Core/Core.h"
 
 namespace Pika {
 
@@ -51,7 +51,7 @@ namespace Pika {
 		template<typename T, typename F> //T:Event£¬F:function
 		bool dispatch(const F& func) {
 			if (m_Event.getEventType() == T::getStaticType()) {
-				m_Event.m_Handled |= func(*(T*)&m_Event);
+				m_Event.m_Handled |= func(*static_cast<T*>(& m_Event));
 				return true;
 			}
 			return false;
