+++ Pika/src/pkpch.h
@@ -13,4 +13,4 @@
 #include <Windows.h>
 
 //Pika engine
-#include "Pika/Log.h"
+#include "Pika/Core/Log.h"
