+++ Sandbox/src/SandboxApp.cpp
@@ -1,10 +1,19 @@
-#include <Pika.h>
-#include <memory>
+#include "Pika.h"
+
+class ExampleLayer : public Pika::Layer
+{
+public:
+	ExampleLayer() :Layer{ "Example Layer" } {};
+	void onUpdate() { PK_INFO("Example Layer update!"); }
+	void onEvent(const Pika::Event& vEvent) { PK_TRACE(vEvent.toString()); }
+};
 
 class Sandbox : public Pika::Application
 {
 public:
-	Sandbox() {};
+	Sandbox() {
+		pushLayer(new ExampleLayer());
+	};
 	~Sandbox() {};
 };
